Add Docker restart policy
When the EC2 instance is rebooted, Artifactory won't start automatically. It will be restarted when Chef runs the cookbook, but it might take up to 15 minutes.

This commits adds a restart policy so that docker restarts it when it crashes or when the ec2 instance boots up (the default restart policy is 'no').

References:
- https://docs.docker.com/engine/admin/host_integration/
- https://docs.docker.com/engine/reference/run/#restart-policies-restart
- https://github.com/chef-cookbooks/docker#properties-6